This commit includes the whole refactored administrator panel version using Redux as a state management container. The new dependences that have been added are redux and redux-thunk.
Also added a "smart way" of fetching/sending data to/from the client in order to reduce the number of calls done on the API. This will reduce the amount of megabytes getting consumed by a client and will reduce the API scalling needs.

The Pages folder and some components that are no longer used due to the redux implementation will be deleted as soon as possible.

All of the changes in this commit have been done to the admin front end section so the commited files of the API section that seem to be modified must have been due to file permissions.